+++ tests/testthat/test-lancelot-support.R
@@ -511,7 +511,7 @@ test_that("Can compute initial conditions", {
   ## * 38 (S + N_tot)
   ## * 1 (react_pos)
   ## * 3 (N_tot_sero_1 + N_tot_sero_2 + N_tot_PCR)
-  expect_equal(sum(initial != 0), 43)
+  expect_equal(sum(initial != 0), 44)
 })
 
 
@@ -539,7 +539,7 @@ test_that("Can control the seeding", {
   ## * 38 (S + N_tot)
   ## * 1 (prob_strain)
   ## * 3 (N_tot_sero_1 + N_tot_sero_2 + N_tot_PCR)
-  expect_equal(sum(initial != 0), 43)
+  expect_equal(sum(initial != 0), 44)
 })
 
 ## TODO: Ed - you had said that you had ideas for some more systematic
@@ -841,7 +841,7 @@ test_that("lancelot_index switches work as expected", {
       info$index$cum_n_vaccinated,
       info$index$D,
       info$index$D_hosp,
-      info$index$infections_inc_per_strain))
+      info$index$infections_inc_strain))
 
   index_severity <- lancelot_index(info, severity = TRUE)
   expect_equal(
@@ -864,19 +864,46 @@ test_that("lancelot_index switches work as expected", {
       info$index$cum_n_vaccinated,
       info$index$D,
       info$index$D_hosp,
-      info$index$infections_inc_per_strain,
+      info$index$infections_inc_strain,
       info$index$ifr,
       info$index$ihr,
       info$index$hfr,
+      info$index$hospitalisations_inc,
       info$index$ifr_strain,
       info$index$ihr_strain,
       info$index$hfr_strain,
+      info$index$hospitalisations_inc_strain,
       info$index$ifr_age,
       info$index$ihr_age,
       info$index$hfr_age,
-      info$index$ifr_disag,
-      info$index$ihr_disag,
-      info$index$hfr_disag))
+      info$index$infections_inc_age,
+      info$index$hospitalisations_inc_age))
+
+  index_protected <- lancelot_index(info, protected = TRUE)
+  expect_equal(
+    unname(index_protected$state[
+      !names(index_protected$state) %in% names(index_protected$run)]),
+    c(info$index$hosp_tot,
+      info$index$cum_admit_conf,
+      info$index$cum_new_conf,
+      info$index$D_tot,
+      info$index$D_inc,
+      info$index$cum_infections,
+      info$index$infections_inc,
+      info$index$effective_susceptible,
+      info$index$S,
+      info$index$R,
+      info$index$prob_strain,
+      info$index$cum_admit_by_age,
+      info$index$diagnoses_admitted,
+      info$index$cum_infections_disag,
+      info$index$cum_n_vaccinated,
+      info$index$D,
+      info$index$D_hosp,
+      info$index$infections_inc_strain,
+      info$index$protected_S_vaccinated,
+      info$index$protected_R_unvaccinated,
+      info$index$protected_R_vaccinated))
 
   index_rt <- lancelot_index(info, rt = FALSE)
   expect_equal(
@@ -895,7 +922,7 @@ test_that("lancelot_index switches work as expected", {
       info$index$cum_n_vaccinated,
       info$index$D,
       info$index$D_hosp,
-      info$index$infections_inc_per_strain))
+      info$index$infections_inc_strain))
 
 
   index_cum_admit <- lancelot_index(info, cum_admit = FALSE)
@@ -918,7 +945,7 @@ test_that("lancelot_index switches work as expected", {
       info$index$cum_n_vaccinated,
       info$index$D,
       info$index$D_hosp,
-      info$index$infections_inc_per_strain))
+      info$index$infections_inc_strain))
 
   index_diagnoses_admitted <- lancelot_index(info, diagnoses_admitted = FALSE)
   expect_equal(
@@ -941,7 +968,7 @@ test_that("lancelot_index switches work as expected", {
       info$index$cum_n_vaccinated,
       info$index$D,
       info$index$D_hosp,
-      info$index$infections_inc_per_strain))
+      info$index$infections_inc_strain))
 
   index_cum_infections_disag <- lancelot_index(info,
    cum_infections_disag = FALSE)
@@ -965,7 +992,7 @@ test_that("lancelot_index switches work as expected", {
       info$index$cum_n_vaccinated,
       info$index$D,
       info$index$D_hosp,
-      info$index$infections_inc_per_strain))
+      info$index$infections_inc_strain))
 
   index_cum_n_vaccinated <- lancelot_index(info, cum_n_vaccinated = FALSE)
   expect_equal(
@@ -988,7 +1015,7 @@ test_that("lancelot_index switches work as expected", {
       info$index$cum_infections_disag,
       info$index$D,
       info$index$D_hosp,
-      info$index$infections_inc_per_strain))
+      info$index$infections_inc_strain))
 
   index_inf_per_strain <-
     lancelot_index(info, infections_inc_per_strain = FALSE)
@@ -1035,7 +1062,7 @@ test_that("lancelot_index switches work as expected", {
       info$index$cum_infections_disag,
       info$index$cum_n_vaccinated,
       info$index$D_hosp,
-      info$index$infections_inc_per_strain))
+      info$index$infections_inc_strain))
 
   index_D_hosp <- lancelot_index(info, D_hosp = FALSE)
   expect_equal(
@@ -1058,7 +1085,7 @@ test_that("lancelot_index switches work as expected", {
       info$index$cum_infections_disag,
       info$index$cum_n_vaccinated,
       info$index$D,
-      info$index$infections_inc_per_strain))
+      info$index$infections_inc_strain))
 })